refactor(review): relocate claim-mapping extraction to TlsClientAuthentication
MtlsClaimsEnhancer's subject CN/OU/O extraction pipeline (RDN parsing via
LdapName, pattern matching) moves to TlsClientAuthentication.extractClaimMappingValues,
since the upcoming tls-client-auth-required-claims enforcement needs it at
authentication time, not just claims-enhancement time. MtlsClaimsEnhancer now
delegates to the shared method instead of duplicating the logic -- pure
extract-and-relocate, no algorithmic change.

MtlsClaimsEnhancerTest switches from a plain mock to a spy of TlsClientAuthentication
so the real (now relocated) extraction still runs during these tests, exactly as
before -- no test body changes needed beyond setUp(), except one stub in
stringPathInAdditionalInformationLoadsTrustedProxyCa() switched from when/thenReturn
to doReturn/when to avoid Mockito's real-method-invocation-during-stubbing behavior
on spies, which was double-counting an interaction verified by that test.
